py/pyproject.toml
[build-system]
build-backend = 'hatchling.build'
requires = ['hatchling']

[project]
classifiers = ['Development Status :: 3 - Alpha', 'Intended Audience :: Developers', 'Topic :: Software Development', 'Typing :: Typed', 'Operating System :: POSIX :: Linux', 'Operating System :: Microsoft :: Windows', 'Operating System :: MacOS', 'Programming Language :: Python', 'Programming Language :: Python :: Implementation :: CPython', 'Programming Language :: Python :: 3', 'Programming Language :: Python :: 3 :: Only', 'Programming Language :: Python :: 3.8', 'Programming Language :: Python :: 3.9', 'Programming Language :: Python :: 3.10', 'Programming Language :: Python :: 3.11', 'Programming Language :: Python :: 3.12', 'License :: OSI Approved']
dependencies = ['pathspec>=0.10.0', 'jinja2>=3.0.0', 'pyyaml>=6.0.0', 'typer>=0.9.0']
description = 'In-place, continuous templater. Powered by Jinja.'
name = 'etcher'
readme = 'README.md'
requires-python = '>=3.8'
version = '0.0.9'

[[project.authors]]
email = 'zakstucke@gmail.com'
name = 'Zachary Stucke'

[project.license]
text = 'MIT'

[project.scripts]
etch = 'etcher.main:cli'

[project.urls]
Documentation = 'https://zakstucke.github.io/etcher'
Sourcecode = 'https://github.com/zakstucke/etcher'

[tool]

[tool.coverage]

[tool.coverage.paths]
source = ['etcher/', '/Users/runner/work/etcher/etcher/etcher/', '/home/runner/work/etcher/etcher/etcher/', 'D:\a\etcher\etcher\etcher']
tests = ['tests']

[tool.coverage.report]
fail_under = 100
show_missing = true

[tool.coverage.run]
source = ['etcher']

[tool.hatch]

[tool.hatch.build]

[tool.hatch.build.targets]

[tool.hatch.build.targets.sdist]
include = ['/README.md', '/LICENSE.md', '/etcher', '/tests', '/requirements.txt']

[tool.hatch.metadata]
allow-direct-references = true

[tool.pdm]

[tool.pdm.dev-dependencies]
dev = ['ipykernel>=6.25.2']
test = ['pyright>=1.1.329', 'pytest>=7.4.2', 'coverage[toml]>=7.3.2']

[tool.pyright]
deprecateTypingAliases = true
exclude = ['**/__pycache__', '**/.pytest_cache', '**/.pytype', '**/py_stubs', '**/venv', '**/.venv', '**/node_modules', '**/target', '**/migrations']
pythonVersion = '3.8'
reportImportCycles = true
reportMissingImports = true
reportMissingTypeStubs = false
reportOptionalCall = true
reportOptionalContextManager = true
reportOptionalIterable = true
reportOptionalMemberAccess = true
reportOptionalOperand = true
reportOptionalSubscript = true
reportPrivateImportUsage = false
reportTypedDictNotRequiredAccess = false
reportUnknownParameterType = true
reportUnnecessaryCast = true
reportUnnecessaryTypeIgnoreComment = false
reportUnusedExpression = false
strictDictionaryInference = true
strictListInference = true
strictSetInference = true
stubPath = './py_stubs'
typeCheckingMode = 'basic'
useLibraryCodeForTypes = true
venv = '.venv'
venvPath = '.'
